# components/chat_client_service_api_client/openapi.json
{"openapi":"3.1.0","info":{"title":"Chat Client Service","description":"Provider-backed chat operations behind the shared ChatClient API.","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","description":"Return service health status.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/auth/sessions":{"post":{"tags":["auth"],"summary":"Create Auth Session","description":"Create a pending service auth session for adapter clients.","operationId":"create_auth_session_auth_sessions_post","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSessionResponse"}}}}}}},"/auth/login/config":{"get":{"tags":["auth"],"summary":"Auth Login Config","description":"Return init data for Telegram.Login JavaScript library.","operationId":"auth_login_config_auth_login_config_get","parameters":[{"name":"session_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","minLength":1},{"type":"null"}],"title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramLoginConfigResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/callback":{"get":{"tags":["auth"],"summary":"Auth Callback","description":"Complete Telegram OIDC login and issue a local Bearer token.","operationId":"auth_callback_auth_callback_get","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}},{"name":"state","in":"query","required":true,"schema":{"type":"string","title":"State"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["auth"],"summary":"Auth Login Library Callback","description":"Complete Telegram.Login JavaScript id_token callback.","operationId":"auth_login_library_callback_auth_callback_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramLoginCallbackRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/telegram-login":{"post":{"tags":["auth"],"summary":"Auth Telegram Hash Callback","description":"Complete Telegram hash login returned as a URL fragment.","operationId":"auth_telegram_hash_callback_auth_telegram_login_post","parameters":[{"name":"telegram_auth_state","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Telegram Auth State"}},{"name":"telegram_auth_session_id","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Telegram Auth Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramHashLoginCallbackRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/verify":{"post":{"tags":["auth"],"summary":"Auth Verify Legacy","description":"Compatibility shim for the parent-branch /auth/verify endpoint.","operationId":"auth_verify_legacy_auth_verify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LegacyTelegramVerifyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OAuthCallbackResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Auth Me","description":"Return identity from the authenticated local session.","operationId":"auth_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}},"security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}]}},"/auth/sessions/{session_id}":{"get":{"tags":["auth"],"summary":"Get Auth Session","description":"Return the current auth state for a service session.","operationId":"get_auth_session_auth_sessions__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSessionStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["auth"],"summary":"Delete Auth Session","description":"Delete a service auth session.","operationId":"delete_auth_session_auth_sessions__session_id__delete","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/messages":{"post":{"tags":["chat"],"summary":"Send Message","description":"Send a message through the configured ChatClient provider.","operationId":"send_message_chat_messages_post","security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["chat"],"summary":"Get Messages","description":"Retrieve messages from a channel through the configured provider.","operationId":"get_messages_chat_messages_get","security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}],"parameters":[{"name":"channel_id","in":"query","required":true,"schema":{"type":"string","title":"Channel Id"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","exclusiveMinimum":0},{"type":"null"}],"title":"Limit"}},{"name":"max_results","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","exclusiveMinimum":0},{"type":"null"}],"title":"Max Results"}},{"name":"cursor","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cursor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageModel"},"title":"Response Get Messages Chat Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/messages/{message_id}":{"get":{"tags":["chat"],"summary":"Get Message","description":"Retrieve a single message by opaque id.","operationId":"get_message_chat_messages__message_id__get","security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}],"parameters":[{"name":"message_id","in":"path","required":true,"schema":{"type":"string","title":"Message Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["chat"],"summary":"Delete Message","description":"Delete a message through the configured ChatClient provider.","operationId":"delete_message_chat_messages__message_id__delete","security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}],"parameters":[{"name":"message_id","in":"path","required":true,"schema":{"type":"string","title":"Message Id"}},{"name":"channel_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Channel Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/channels":{"get":{"tags":["chat"],"summary":"Get Channels","description":"List available channels through the configured ChatClient provider.","operationId":"get_channels_chat_channels_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ChannelModel"},"type":"array","title":"Response Get Channels Chat Channels Get"}}}}},"security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}]}},"/chat/channels/{channel_id}":{"get":{"tags":["chat"],"summary":"Get Channel","description":"Retrieve a single known channel by id.","operationId":"get_channel_chat_channels__channel_id__get","security":[{"HTTPBearer":[]},{"APIKeyHeader":[]},{"APIKeyCookie":[]}],"parameters":[{"name":"channel_id","in":"path","required":true,"schema":{"type":"string","title":"Channel Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/telegram/webhook":{"post":{"tags":["telegram"],"summary":"Telegram Webhook","description":"Record supported Telegram Bot API updates for later read endpoints.","operationId":"telegram_webhook_telegram_webhook_post","parameters":[{"name":"X-Telegram-Bot-Api-Secret-Token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Telegram-Bot-Api-Secret-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramUpdate"}}}},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthSessionResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"authenticated":{"type":"boolean","title":"Authenticated"},"login_url":{"type":"string","title":"Login Url"},"status_url":{"type":"string","title":"Status Url"},"bot_username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bot Username"},"bot_start_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bot Start Url"}},"type":"object","required":["session_id","authenticated","login_url","status_url"],"title":"AuthSessionResponse","description":"Pending service auth session response."},"AuthSessionStatusResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"authenticated":{"type":"boolean","title":"Authenticated"},"telegram_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Telegram Id"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","required":["session_id","authenticated"],"title":"AuthSessionStatusResponse","description":"Current service auth session state."},"ChannelModel":{"properties":{"id":{"type":"string","minLength":1,"title":"Id"},"name":{"type":"string","minLength":1,"title":"Name"},"channel_type":{"type":"string","minLength":1,"title":"Channel Type"}},"type":"object","required":["id","name","channel_type"],"title":"ChannelModel","description":"HTTP model for a provider-backed chat channel."},"DeleteMessageResponse":{"properties":{"success":{"type":"boolean","title":"Success"}},"type":"object","required":["success"],"title":"DeleteMessageResponse","description":"Response body for deleting a message."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"HealthResponse","description":"Health endpoint response."},"LegacyTelegramVerifyRequest":{"properties":{"id":{"type":"integer","title":"Id"},"first_name":{"type":"string","title":"First Name"},"auth_date":{"type":"integer","title":"Auth Date"},"hash":{"type":"string","minLength":1,"title":"Hash"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"photo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Url"}},"type":"object","required":["id","first_name","auth_date","hash"],"title":"LegacyTelegramVerifyRequest","description":"Decoded Telegram login payload used by the legacy /auth/verify route."},"LogoutResponse":{"properties":{"success":{"type":"boolean","title":"Success"}},"type":"object","required":["success"],"title":"LogoutResponse","description":"Service auth session deletion response."},"MeResponse":{"properties":{"telegram_id":{"type":"string","title":"Telegram Id"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","required":["telegram_id"],"title":"MeResponse","description":"Authenticated Telegram identity extracted from the service token."},"MessageModel":{"properties":{"id":{"type":"string","minLength":1,"title":"Id"},"sender":{"type":"string","title":"Sender"},"channel_id":{"type":"string","minLength":1,"title":"Channel Id"},"timestamp":{"type":"string","title":"Timestamp"},"text":{"type":"string","title":"Text"}},"type":"object","required":["id","sender","channel_id","timestamp","text"],"title":"MessageModel","description":"HTTP model for a provider-backed chat message."},"OAuthCallbackResponse":{"properties":{"detail":{"type":"string","title":"Detail"},"code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"access_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"},"token_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Token Type"}},"type":"object","required":["detail"],"title":"OAuthCallbackResponse","description":"Legacy response shape returned by /auth/verify."},"SendMessageRequest":{"properties":{"channel_id":{"type":"string","minLength":1,"title":"Channel Id"},"text":{"type":"string","maxLength":4096,"minLength":1,"title":"Text"}},"type":"object","required":["channel_id","text"],"title":"SendMessageRequest","description":"Request body for sending a message."},"TelegramHashLoginCallbackRequest":{"properties":{"auth_result":{"type":"string","minLength":1,"title":"Auth Result"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id"}},"type":"object","required":["auth_result"],"title":"TelegramHashLoginCallbackRequest","description":"Signed Telegram login payload returned in tgAuthResult fragments."},"TelegramLoginCallbackRequest":{"properties":{"id_token":{"type":"string","minLength":1,"title":"Id Token"},"nonce":{"type":"string","minLength":1,"title":"Nonce"}},"type":"object","required":["id_token","nonce"],"title":"TelegramLoginCallbackRequest","description":"ID token callback payload from Telegram.Login JavaScript."},"TelegramLoginConfigResponse":{"properties":{"client_id":{"type":"string","title":"Client Id"},"nonce":{"type":"string","title":"Nonce"},"origin":{"type":"string","title":"Origin"},"bot_username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bot Username"},"bot_start_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bot Start Url"}},"type":"object","required":["client_id","nonce","origin"],"title":"TelegramLoginConfigResponse","description":"Configuration for the Telegram Login JavaScript library."},"TelegramUpdate":{"properties":{"update_id":{"type":"integer","title":"Update Id"}},"additionalProperties":true,"type":"object","required":["update_id"],"title":"TelegramUpdate","description":"Permissive model for Telegram Bot API webhook updates."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse","description":"Bearer token issued by this service after Telegram OIDC login."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"},"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-Session-ID"},"APIKeyCookie":{"type":"apiKey","in":"cookie","name":"chat_client_session"}}}}